==> getting_started_with_ml/notebooks/0_intro_notebooks.ipynb <==
# Intro to Python and Notebooks

## Python

Python is an intepreted language, that means it executes its code directly as it is written by translating it into specific machine instructions every time it is run. A change to your code is immediately effective upon its next execution. This is different from so called compiler languages which have to have their code translated (compiled) into an executable program that is then called and needs to be recompiled if changes to the original code are made.  

The great benefit of interpreted languages is that you can write code, quickly execute it and change it if you want a different result. 

## Notebooks

### Cells

- In notebooks, we have cells that hold code blocks that get executed together.
- In Jupyter:
    - You can run a cell using __Shift-Enter__.
    - You can insert cells using __Insert__ in the menu-bar.
    - You can delete cells using __d + d__ (selecting the chunk and pressing `d` twice). 
- In Google Colab:
    - You can run a cell using the __Play__ button.
    - You can insert cells using __Insert__ in the menu-bar.
    - You can delete cells using __Right Click > Delete Cell__. 

### Execution

- You can execute these code blocks in any order you like, and the number in the square brackets to the left of the block shows at which point in your series of executions this cell was executed. 
- For example, if you execute your first cell (they count only code, not these comments that are written in a language called "markdown"), the number 1 appears in the brackets next to the cell and, if the code has an output, also in the brackets next to the output. 
- If you execute two cells consecutively, their numbers will be 1 apart, increasing in order of execution. 
- Be careful and keep track of your order, else you might have old settings, variables, functions or other junk from old executions in memory that are still in effect even though you deleted or changed their cells after execution.  

- There are a lot more details about Jupyter lab and Jupyter notebooks in the reference, which you can access from the "Help" section in the menu bar.

## Python Basics

x = 3

y = 4

By default, the notebook will print the value of the last value called in the cell: 

y
x

But we can also explicitly print variables:

print(y)
print(x)

Maths operations are mostly as you would expect:

x + x

x * x

# Many math operations are available as functions
pow(x,x)

Printing strings:

print("hello world")

Defining a function in python:

def my_func(x, y):
    return (x ** x) + (y ** y)

And then calling it:

my_func(3, 5)

Finally, we can also define functions within functions, like a closure:

def my_func_2(x, y):
    def other_func(z):
        return x * y * z
    
    return other_func

When you call this function, it doesn't immediately produce a result like a value assigned to a variable - it returns a *function*.  

Try calling `my_func_2` like any other variable and you'll see that Python returns an object instead of a value.

my_func_2

Now check this out:

f = my_func_2(3, 5)
print(f(4))
print(f(8))
